Struct rsnake::game::Game

pub struct Game {
    snake: Snake,
    fruit: Position,
    size: (u32, u32),
    waiting_time: f64,
    score: u32,
    over: bool,
    paused: bool,
    wrote: bool,
    supfruit: Position,
}

Fields
snake: Snakefruit: Positionsize: (u32, u32)waiting_time: f64score: u32over: boolpaused: boolwrote: boolsupfruit: Position
Implementations

impl Game

pub fn new(width: u32, height: u32) -> Self

this is the starting values of everything present in the game

pub fn start(&mut self)

starts the game back up if paused, can do so by pressing any key.

pub fn pause(&mut self)

pauses the game, can do so with the key R

pub fn toggle_game_state(&mut self)

changes the paused state to either resume or pause the game

pub fn draw(&mut self, ctx: Context, g: &mut G2d<'_>)

added code that prints the high scores and if the user gets a new high score it’s added to the top five and the smallest score is deleted

pub fn update(&mut self, delta_time: f64)

Updates game information such as the snake length and score when picking up a fruit Added a blue fruit that gives more points and segments

pub fn key_down(&mut self, key: Key)

takes key presses from the user uses them to move the snake or pause and restart the game.

pub fn get_score(&self) -> u32

returns the game score

fn calc_score(&mut self)

calculates the score based on the length of the snake times 10

fn restart(&mut self)

This function restarts the game