rsnake

Module graphics

Game graphical Modules

This module contains the core components of the game logic, it includes the following submodules:

graphic_block: Handles the graphical rendering of game logic blocks in the terminal.

sprites: Provides the game sprites, as snake, map and so on. .

menus: Provides utility functions to assist with common texts and graphical tasks

Modules

graphic_block

menus

This module provides various menu and UI components for the game interface.

playing_render

sprites

This module contains the core components of the game logic, it includes the following sub-modules: