Struct rsnake::snake::Snake

pub struct Snake {
    direction: Direction,
    head: Position,
    tail: LinkedList<Position>,
    updated_tail_pos: bool,
}

Fields
direction: Directionhead: Positiontail: LinkedList<Position>updated_tail_pos: bool
Implementations

impl Snake

pub fn new(head: Position) -> Self

creates a new snake

pub fn update(&mut self, width: u32, height: u32)

updates the snakes direction, length and position.

pub fn draw(&self, ctx: &Context, g: &mut G2d<'_>)

draws the snakes head and its tail

pub fn set_dir(&mut self, dir: Direction)

sets the direction the snake is pointing

pub fn get_head_pos(&self) -> &Position

returns the position of the snakes head

pub fn get_len(&self) -> usize

returns the length of the tail

pub fn is_tail_overlapping(&self) -> bool

returns a bool showing that the head and the tail are currently overlapping

pub fn will_tail_overlapp(&self) -> bool

returns a bool showing whether the head and the tail will overlap next position

pub fn grow(&mut self)

Increases the length of the snake ehwn called

fn next_head_pos(&self) -> Position

returns the next position of the head